% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forwardBackwardHMMc}
\alias{forwardBackwardHMMc}
\title{Forward-backward algorithm implemented in Rcpp}
\usage{
forwardBackwardHMMc(obs, transMtx, emisMtx, pi)
}
\arguments{
\item{obs}{size T integer vector of observed outcomes, values from 1 to m}

\item{transMtx}{n x n transition matrix from row to column}

\item{emisMtx}{n x m emission matrix of observed data (column) given states (row)}

\item{pi}{size n vector of initial state probabilities}
}
\value{
matrix of conditional probability of each state given the observed outcomes
}
\description{
This function implements a forward-backward algorithm of a generic HMM, given observed outcomes, transition and emission matrices, and initial probabilities.
}
\examples{
obs <- c(1,3,2)
A <- matrix(c(0.8,0.2,0.4,0.6),2,2,byrow=TRUE)
B <- matrix(c(0.88,0.10,0.02,0.1,0.6,0.3),2,3,byrow=TRUE)
pi <- c(0.7,0.3)
forwardBackwardHMMc(obs, A, B, pi)
}
