% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmr.r
\name{forwardBackwardHMMr}
\alias{forwardBackwardHMMr}
\title{Forward-backward algorithm implemented in R}
\usage{
forwardBackwardHMMr(obs, transMtx, emisMtx, pi)
}
\arguments{
\item{obs}{size T integer vector of observed outcomes, values from 1 to m}

\item{transMtx}{n x n transition matrix from row to column}

\item{emisMtx}{n x m emission matrix of observed data (column) given states (row)}

\item{pi}{size n vector of initial state probabilities}
}
\value{
a matrix containing conditional probability of each possible states given the observed outcomes
}
\description{
This function implements a forward-backward algorithm for a generic HMM in R, given observed outcomes, transition and emission matrices, and initial probabilities.
}
\examples{
obs <- c(1,3,2) 
A <- matrix(c(0.8,0.2,0.4,0.6),2,2,byrow=TRUE)
B <- matrix(c(0.88,0.10,0.02,0.1,0.6,0.3),2,3,byrow=TRUE)
pi <- c(0.7,0.3)
forwardBackwardHMMr(obs, A, B, pi)

}
